S09P22A708-519 Fix: redis 설정 에러 해결

+++ jazz_backend/src/main/java/com/ssafy/jazz_backend/domain/member/record/repository/SeasonJpaRepository.java
@@ -3,6 +3,6 @@ package com.ssafy.jazz_backend.domain.member.record.repository;
 import com.ssafy.jazz_backend.domain.member.record.entity.Season;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeasonJpaRepository extends JpaRepository<Season, Long> {
+public interface SeasonJpaRepository extends JpaRepository<Season, Integer> {
 
 }
